some changes

+++ src/app/layout/left-bar/left-bar.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common'; // Add this line
 import { ActivatedRoute, Router } from '@angular/router';
 import { environment } from 'src/environments/environment';
+import Swal from 'sweetalert2';
 const clientName = `${environment.default}`;
 @Component({
   standalone: true,
@@ -84,8 +85,35 @@ export class LeftBarComponent {
 
   }
   clearLocalStorage() {
-    localStorage.clear();
-    this.router.navigate([clientName + '/sign-in'])
+    Swal.fire({
+      title: 'Confirmer les modifications',
+      html: `
+        <div>
+        <div style="font-size:1.2rem;">  Êtes-vous sûr de vouloir <br>déconnecter de ce compte ? </div> 
+        </div>
+      `,
+      iconColor: '#1E1E1E',
+      background: '#fefcf1',
+      showCancelButton: true,
+      confirmButtonText: 'Oui',
+      confirmButtonColor: "#91c593",
+      cancelButtonText: 'Non',
+      cancelButtonColor: "black",
+      customClass: {
+        confirmButton: 'custom-confirm-button-class',
+        cancelButton: 'custom-cancel-button-class'
+      },
+      reverseButtons: true // Reversing button order
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        localStorage.clear();
+        this.router.navigate([clientName + '/sign-in'])
+        // User clicked 'Yes', call the endpoint
+
+      }
+    });
+
   }
 }
